pyproject.toml
[tool.poetry]
name = "llm-test"
version = "0.1.0"
description = "LLM Test Project"
authors = ["tony@kormedi.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
pandas = "^2.2.2"
pymysql = "^1.1.1"
python-dotenv = "^1.0.1"
langchain = "0.2.17"
langchain-core = "0.2.43"
langchain-openai = "0.1.25"
langchain-community = "0.2.19"
langchain-aws = "0.1.18"
langchain-ollama = "0.1.3"
aiohttp = "^3.11.14"
boto3 = "^1.34.0"
numpy = "^1.26.4"
openai = "^1.43.0"
pydantic = "2.8.2"
sqlalchemy = "^2.0.32"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
